requirements.txt
requests==2.23.0
pymongo==3.10.1
uplink==0.9.1
dacite==1.3.0
Unidecode==1.1.1
mongo-thingy==0.12.0
requests_oauthlib==1.3.0
dependency_injector==3.15.6
PyYAML==5.3.1
asyncio==3.4.3
pytz==2019.3